refactor(client): extract modal from table
