+++ client/src/components/views/Orders.jsx
@@ -1,57 +1,17 @@
-import React, { Fragment, useState, useRef } from 'react';
-import {
-  NumberInput,
-  NumberInputField,
-  NumberInputStepper,
-  NumberIncrementStepper,
-  NumberDecrementStepper,
-  Center,
-  useToast,
-  Textarea,
-  Select,
-  useDisclosure,
-  Table,
-  Thead,
-  Tbody,
-  Tr,
-  TableCaption,
-  Heading,
-  Button,
-  FormControl,
-  FormLabel,
-  Input,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalFooter,
-  ModalBody,
-  ModalCloseButton,
-} from '@chakra-ui/react';
-import { SmallAddIcon } from '@chakra-ui/icons';
-import Loader from 'components/common/Loader.jsx';
-import { getOrders, getMaterialList, postOrders } from 'utils/api/orders.js';
-import { useQuery } from 'react-query';
-import { StyledTableRow, StyledTableHeader, StyledTableCell } from 'components/common/Table.jsx';
+import { Heading, Table, TableCaption, Tbody, Thead, Tr } from '@chakra-ui/react';
 import { TablePagination } from '@material-ui/core';
 import { NoResultImage } from 'components/common/Image.jsx';
+import Loader from 'components/common/Loader.jsx';
+import { StyledTableCell, StyledTableHeader, StyledTableRow } from 'components/common/Table.jsx';
+import FormModal from 'components/Order/FormModal.jsx';
+import React, { Fragment, useState } from 'react';
+import { useQuery } from 'react-query';
+import { getOrders } from 'utils/api/orders.js';
 
 const Orders = () => {
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const { isLoading, isSuccess, data, refetch } = useQuery('orders', getOrders);
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
-  const toast = useToast();
-  const [material, setMaterial] = useState();
-  const [cost, setCost] = useState(0);
-  const [deliveryDate, setDeliveryDate] = useState();
-  const [orderDate, setOrderDate] = useState();
-  const [quantity, setQuantity] = useState(0);
-  const manufacturer = useRef('');
-  const location = useRef('');
-  const note = useRef('');
-  const materialTypes = ['rubber', 'aluminum', 'steel', 'copper', 'plastic', 'leather'];
-  const materialCost = useQuery('orders/materialList', getMaterialList);
+  const { isLoading, isSuccess, data } = useQuery('orders', getOrders);
 
   const handleChangePage = (event, newPage) => {
     setPage(newPage);
@@ -61,109 +21,6 @@ const Orders = () => {
     setRowsPerPage(parseInt(event.target.value, 10));
     setPage(0);
   };
-
-  // sets material and cost states
-  const handleMaterial = (e) => {
-    const choice = e.target.value;
-    setMaterial(choice);
-    setCost(materialCost.data.data[choice].cost);
-  };
-
-  // generates a random int within ranges
-  const getRandomInt = (min, max) => {
-    return Math.floor(Math.random() * (max - min)) + min;
-  };
-
-  const handleSubmit = async () => {
-    const orderTime = new Date();
-    const deliveryTime = new Date();
-    const randomDeliveryDate = deliveryTime.getDate() + getRandomInt(1, 8); // random delivery date within a week
-    deliveryTime.setDate(randomDeliveryDate);
-    setOrderDate(orderTime);
-    setDeliveryDate(deliveryTime);
-
-    try {
-      await postOrders({
-        materialType: material,
-        cost: cost,
-        quantity: quantity,
-        deliveryDate: deliveryDate,
-        orderDate: orderDate,
-        manufacturerName: manufacturer.current.value,
-        vendorLocation: location.current.value,
-        status: 'Pending',
-        note: note.current.value,
-      });
-      toast({
-        title: 'Order Placed',
-        description: 'The order has been placed',
-        status: 'success',
-        duration: 9000,
-        isClosable: true,
-      });
-    } catch {
-      toast({
-        position: 'top',
-        title: 'An error occurred.',
-        description: 'Unable to place order.',
-        status: 'error',
-        duration: 2000,
-        isClosable: true,
-      });
-    }
-    refetch();
-    onClose();
-  };
-
-  const openOrderForm = () => (
-    <Modal isOpen={isOpen} onClose={onClose}>
-      <ModalOverlay />
-      <ModalContent>
-        <ModalHeader>Place an order</ModalHeader>
-        <ModalCloseButton />
-        <ModalBody pb={6}>
-          <FormControl isRequired>
-            <FormLabel>Material</FormLabel>
-            <Select placeholder="Select an option" onChange={handleMaterial}>
-              {materialTypes.map((material) => (
-                <Fragment key={material._id}>
-                  <option value={material}>{material}</option>
-                </Fragment>
-              ))}
-            </Select>
-          </FormControl>
-          <FormControl isRequired>
-            <FormLabel>Quantity</FormLabel>
-            <NumberInput min={1} defaultValue={1} onChange={(e) => setQuantity(e)}>
-              <NumberInputField />
-              <NumberInputStepper>
-                <NumberIncrementStepper />
-                <NumberDecrementStepper />
-              </NumberInputStepper>
-            </NumberInput>
-          </FormControl>
-          <FormControl mt={4}>
-            <FormLabel>Manufacturer</FormLabel>
-            <Input placeholder="Manufacturer" ref={manufacturer} />
-          </FormControl>
-          <FormControl mt={4}>
-            <FormLabel>Location</FormLabel>
-            <Input placeholder="Location" ref={location} />
-          </FormControl>
-          <FormControl mt={4}>
-            <FormLabel>Note to the manufacturer</FormLabel>
-            <Textarea placeholder="Note" ref={note} />
-          </FormControl>
-        </ModalBody>
-        <ModalFooter>
-          <Button colorScheme="blue" mr={3} onClick={handleSubmit}>
-            Submit
-          </Button>
-          <Button onClick={onClose}>Cancel</Button>
-        </ModalFooter>
-      </ModalContent>
-    </Modal>
-  );
 
   if (isLoading) {
     return <Loader />;
@@ -175,10 +32,7 @@ const Orders = () => {
         <Heading size="xl" textAlign="center" mt={5}>
           No orders found.
         </Heading>
-        <Center mt={4}>
-          <Button onClick={onOpen}>Place an order</Button>
-        </Center>
-        {openOrderForm()}
+        <FormModal showButton={true} />
         <NoResultImage />
       </>
     );
@@ -193,16 +47,7 @@ const Orders = () => {
 
   return (
     <>
-      <Button
-        margin={1}
-        colorScheme="blue"
-        variant="solid"
-        leftIcon={<SmallAddIcon />}
-        onClick={onOpen}
-      >
-        Place an order
-      </Button>
-      {openOrderForm()}
+      <FormModal />
       <Table minWidth="unset" width="100%" variant="striped" colorScheme="light">
         <TableCaption placement="top">List of orders</TableCaption>
         <Thead>
